Create array of file objects to append to, since we cannot append to the file input.
File inputs are read-only, so if the user selects files to upload, but then clicks the button to choose some more, the first ones are replaced instead of appended to. I am therefore creating the file object array to append them to so I can send this information along with FormData.